+++ com.unity.visualeffectgraph/Editor/Inspector/VFXContextEditor.cs
@@ -256,6 +256,10 @@ class VFXContextEditor : VFXSlotContainerEditor
         }
     }
 
+    public virtual void DisplayWarnings()
+    {
+    }
+
     public override void OnInspectorGUI()
     {
         if (dataObject != null)
@@ -277,7 +281,7 @@ class VFXContextEditor : VFXSlotContainerEditor
                 ctx.GetData().Invalidate(VFXModel.InvalidationCause.kSettingChanged); // This will also invalidate contexts
             }
         }
-
+        DisplayWarnings();
         DisplaySummary();
     }
 }